% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_norm}
\alias{check_norm}
\title{Check Normalization Procedures}
\usage{
check_norm(.data)
}
\arguments{
\item{.data}{A data frame produced by a tidynorm function.}
}
\description{
\code{check_norm()} will generate a message with information
about which normalization procedures have been applied to the
data.
}
