% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct.R
\name{dct_basis}
\alias{dct_basis}
\title{DCT Basis}
\usage{
dct_basis(n, k)
}
\arguments{
\item{n}{The length of the basis.}

\item{k}{The number of basis functions.}
}
\value{
A \eqn{n\times k} matrix
}
\description{
The Discrete Cosine Transform basis functions
}
\details{
This function will generate the DCT basis functions.
}
\examples{
basis <- dct_basis(100, 5)
matplot(basis, type = "l", lty = 1)

}
