% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct.R
\name{idct}
\alias{idct}
\title{Inverse Discrete Cosine Transform}
\usage{
idct(y, n = length(y))
}
\arguments{
\item{y}{DCT coefficients}

\item{n}{The desired length of the idct}
}
\value{
A vector with the inverse DCT values
}
\description{
The Inverse DCT
}
\details{
See the \link{dct}.
}
\examples{
x <- seq(0,1, length = 10)
y <- 5 + x + (2 * (x^2)) + (-2 * (x^4))

dct_coefs <- dct(y)
recovered_y <- idct(dct_coefs)

plot(y, recovered_y)

}
