% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{identify_outliers}
\alias{identify_outliers}
\title{Identify Outliers}
\usage{
identify_outliers(
  .data,
  ...,
  .by = NULL,
  .pre_trans = function(x) x,
  .level = 0.99,
  .call = caller_env(),
  .silent = opt("tidynorm.silent")
)
}
\arguments{
\item{.data}{A data frame containing vowel formant data}

\item{...}{\code{\link[dplyr:dplyr_tidy_select]{<tidy-select>}} One or more unquoted
expressions separated by commas. These should target the vowel formant
data columns.}

\item{.by}{\code{\link[dplyr:dplyr_tidy_select]{<tidy-select>}} A selection of
columns to group by. Typically a column of speaker IDs.}

\item{.pre_trans}{A function to apply to formant values before normalization.}

\item{.call}{Used for internal purposes.}

\item{.silent}{Suppress normalization information messages when running a \verb{norm_*()} function. (Defaults to \code{FALSE}, overwritable using option 'tidynorm.silent' or environment variable 'R_TIDYNORM_TIDYNORM_SILENT')}
}
\description{
Identify Outliers
}
