% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm_barkz}
\alias{norm_barkz}
\title{Bark Difference Normalization}
\usage{
norm_barkz(
  .data,
  ...,
  .by = TRUE,
  .drop_orig = FALSE,
  .keep_params = FALSE,
  .names = "{.formant}_bz",
  .silent = FALSE
)
}
\description{
Bark Difference Normalization
}
