% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_norm.R
\name{norm_dct_generic}
\alias{norm_dct_generic}
\title{Normalize Formant Track}
\usage{
norm_dct_generic(
  .data,
  ...,
  .token_id_col,
  .by = NULL,
  .param_col = NULL,
  .L = 0,
  .S = 1/sqrt(2),
  .by_formant = FALSE,
  .names = "{.formant}_n",
  .silent = FALSE,
  .drop_orig = FALSE,
  .call = caller_env()
)
}
\description{
Normalize Formant Track
}
