% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm_deltaF}
\alias{norm_deltaF}
\title{Delta F Method}
\usage{
norm_deltaF(
  .data,
  ...,
  .by = NULL,
  .drop_orig = FALSE,
  .keep_params = FALSE,
  .names = "{.col}_df"
)
}
\description{
Delta F Method
}
