% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm_generic}
\alias{norm_generic}
\title{Generic normalization procedure}
\usage{
norm_generic(
  .data,
  ...,
  .by = NULL,
  .norm_fun = identity_norm_fun,
  .by_formant = FALSE,
  .drop_orig = FALSE,
  .keep_params = FALSE,
  .names = "{.col}_n",
  call = caller_env()
)
}
\arguments{
\item{.data}{A data frame containing vowel formant data}

\item{...}{\code{\link[dplyr:dplyr_tidy_select]{<tidy-select>}} One or more unquoted
expressions separated by commas. These should target the vowel formant
data columns.}

\item{.by}{\code{\link[dplyr:dplyr_tidy_select]{<tidy-select>}} A selection of
columns to group by. Typically a column of speaker IDs.}

\item{.norm_fun}{A function to apply to a pivoted long data frame. See
Detsils for how to write such a function.}

\item{.by_formant}{Whether or not the normalization method is formant
intrinsic.}

\item{.drop_orig}{Whether or not to drop the original formant data columns.}

\item{.keep_params}{Whether or not to keep the Location (\verb{*_L}) and scale
(\verb{*_S}) normalization parameters}

\item{.names}{A \code{\link[glue:glue]{glue::glue()}} expression for naming the normalized
data columns. The \code{"{.col}"} portion corresponds to the name of the original
formant columns.}
}
\description{
This is a generic normalization procedure with which
you can create your own normalization method.
}
\details{
The following \verb{norm_*} procedures are built on top of \code{norm_generic()}.
}
