% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm_lobanov}
\alias{norm_lobanov}
\title{Lobanov Normalize}
\usage{
norm_lobanov(
  .data,
  ...,
  .by = NULL,
  .drop_orig = FALSE,
  .keep_params = FALSE,
  .names = "{.col}_z"
)
}
\description{
Lobanov Normalize
}
