% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm_nearey}
\alias{norm_nearey}
\title{Nearey Normalization}
\usage{
norm_nearey(
  .data,
  ...,
  .by = NULL,
  .drop_orig = FALSE,
  .keep_params = FALSE,
  .names = "{.col}_lm"
)
}
\description{
Nearey Normalization
}
