% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_norm.R
\name{norm_track_generic}
\alias{norm_track_generic}
\title{Normalize Formant Track}
\usage{
norm_track_generic(
  .data,
  ...,
  .token_id_col,
  .by = NULL,
  .time_col = NULL,
  .norm_fun = lobanov_dct_norm_fun,
  .by_formant = FALSE,
  .order = 5,
  .return_dct = FALSE,
  .drop_orig = FALSE,
  .keep_params = FALSE,
  .names = "{.col}_n",
  .silent = FALSE,
  .call = caller_env()
)
}
\description{
Normalize Formant Track
}
