% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_norm.R
\name{norm_track_wattfab}
\alias{norm_track_wattfab}
\title{Watt and Fabricius Track normalization}
\usage{
norm_track_wattfab(
  .data,
  ...,
  .token_id_col,
  .by = NULL,
  .time_col = NULL,
  .order = 5,
  .return_dct = FALSE,
  .drop_orig = FALSE,
  .names = "{.formant}_wf",
  .silent = FALSE
)
}
\arguments{
\item{.data}{A data frame containing vowel formant data}

\item{...}{\code{\link[dplyr:dplyr_tidy_select]{<tidy-select>}} One or more unquoted
expressions separated by commas. These should target the vowel formant
data columns.}

\item{.token_id_col}{\code{\link[rlang:args_data_masking]{<data-masking>}} A column
that identifies token ids.}

\item{.by}{\code{\link[dplyr:dplyr_tidy_select]{<tidy-select>}} A selection of
columns to group by. Typically a column of speaker IDs.}

\item{.time_col}{\code{\link[rlang:args_data_masking]{<data-masking>}} A time column.
(optional)}

\item{.order}{The number of DCT parameters to use.}

\item{.return_dct}{Whether or not the normalized DCT coefficients themselves
should be returned.}

\item{.drop_orig}{Should the originally targeted columns be dropped.}

\item{.names}{A \code{\link[glue:glue]{glue::glue()}} expression for naming the normalized
data columns. The \code{"{.formant}"} portion corresponds to the name of the original
formant columns.}

\item{.silent}{Whether or not the informational message should be printed.}
}
\value{
A data frame of Watt & Fabricius normalized formant tracks.
}
\description{
Watt and Fabricius Track normalization
}
\details{
This is a modified version of the Watt & Fabricius Method. The original
method identified point vowels over which F1 and F2 centroids were calculated.
The procedure here just identifies centroids by taking the mean of
all formant values.

\deqn{
\hat{F}_{ij} = \frac{F_{ij}}{S_i}
}

\deqn{
 S_i = \frac{1}{N}\sum_{j=1}^N F_{ij}
}

Where
\itemize{
\item \eqn{\hat{F}} is the normalized formant
\item \eqn{i} is the formant number
\item \eqn{j} is the token number
}
}
\references{
Watt, D., & Fabricius, A. (2002). Evaluation of a technique for improving the
mapping of multiple speakers’ vowel spaces in the F1 ~ F2 plane.
Leeds Working Papers in Linguistics and Phonetics, 9, 159–173.
}
