% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options}
\alias{options}
\title{tidynorm Options}
\description{
Internally used, package-specific options. All options will prioritize R options() values, and fall back to environment variables if undefined. If neither the option nor the environment variable is set, a default value is used.
}
\section{Checking Option Values}{

Option values specific to \code{tidynorm} can be
accessed by passing the package name to \code{env}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{options::opts(env = "tidynorm")

options::opt(x, default, env = "tidynorm")
}\if{html}{\out{</div>}}
}

\section{Options}{

\describe{
\item{tidynorm.silent}{\describe{
Suppress normalization information messages when running a norm_*() function.\item{default: }{\preformatted{FALSE}}
\item{option: }{tidynorm.silent}
\item{envvar: }{R_TIDYNORM_TIDYNORM_SILENT (evaluated if possible, raw string otherwise)}
}}

}
}

\seealso{
options getOption Sys.setenv Sys.getenv
}
