% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct.R
\name{reframe_dct}
\alias{reframe_dct}
\title{Reframe DCT}
\usage{
reframe_dct(
  .data,
  ...,
  .token_id_col,
  .by = NULL,
  .time_col = NULL,
  .order = 5
)
}
\arguments{
\item{.data}{A data frame}

\item{...}{\code{\link[dplyr:dplyr_tidy_select]{<tidy-select>}} One or more unquoted
expressions separated by commas. These should target the vowel formant}

\item{.token_id_col}{A token ID column}

\item{.time_col}{A time column}

\item{.order}{The DCT order}
}
\description{
Reframe DCT
}
