% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct.R
\name{reframe_with_idct}
\alias{reframe_with_idct}
\title{Reframe with idct}
\usage{
reframe_with_idct(
  .data,
  ...,
  .token_id_col,
  .by = NULL,
  .param_col = NULL,
  .n = 20
)
}
\arguments{
\item{.data}{A data frame}

\item{...}{\code{\link[dplyr:dplyr_tidy_select]{<tidy-select>}} One or more unquoted
expressions separated by commas. These should target the vowel formant.}

\item{.token_id_col}{The token ID column.}
}
\description{
Reframe with idct
}
