% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{speaker_tracks}
\alias{speaker_tracks}
\title{Speaker Tracks}
\format{
\subsection{\code{speaker_tracks}}{

A data frame with 20,000 rows and 9 columns
\describe{
\item{speaker}{Speaker ID column}
\item{id}{Within speaker id for each token}
\item{vowel}{CMU Dictionary vowel class}
\item{plt_vclass}{Modified Labov-Trager vowel class}
\item{ipa_vclas}{IPA-like vowel class}
\item{word}{Word that the vowel appeared in}
\item{t}{Measurement time point}
\item{F1, F2, F3}{The first, second and third formants, in Hz}
}
}
}
\usage{
speaker_tracks
}
\description{
Speaker Tracks
}
\keyword{datasets}
