% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidynorm-package.R
\docType{package}
\name{tidynorm-package}
\alias{tidynorm}
\alias{tidynorm-package}
\title{tidynorm: Tools for Tidy Vowel Normalization}
\description{
Provides tools for doing tidy speaker vowel normalization.
}
\author{
\strong{Maintainer}: Josef Fruehwald \email{JoFrhwld@gmail.com}

}
\keyword{internal}
