% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{tidynorm_options}
\alias{tidynorm_options}
\title{Set tidynorm options}
\usage{
tidynorm_options(
  .silent = opt("tidynorm.silent"),
  .warnings = opt("tidynorm.warnings")
)
}
\arguments{
\item{.silent}{Suppress normalization information messages when running a \verb{norm_*()} function. (Defaults to \code{FALSE}, overwritable using option 'tidynorm.silent' or environment variable 'R_TIDYNORM_TIDYNORM_SILENT')}

\item{.warnings}{Print warnings from tidynorm functions. (Defaults to \code{TRUE}, overwritable using option 'tidynorm.warnings' or environment variable 'R_TIDYNORM_TIDYNORM_WARNINGS')}
}
\description{
Set tidynorm verbosity
}
\examples{
tidynorm_options(.silent = TRUE, .warnings = FALSE)

speaker_data |>
  norm_generic(F1:F3) ->
  norm1

tidynorm_options(.silent = FALSE, .warnings = TRUE)

speaker_data |>
  norm_generic(F1:F3) ->
  norm2

}
\seealso{
\link{options}
}
