% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct.R
\name{idct}
\alias{idct}
\title{Inverse Discrete Cosine Transform}
\usage{
idct(y, n)
}
\arguments{
\item{y}{DCT coefficients}

\item{n}{The desired length of the idct}
}
\value{
A vector with the inverse DCT values
}
\description{
The Inverse DCT
}
\details{
Applies the Inverse DCT (see \link{dct} for more details).

\deqn{
x_j = \sqrt{2}y_0 + 2\sum_{k=1}^{N-1} y_k \cos\left(\frac{\pi k(2j+1)}{2J}\right)
}
}
\examples{
x <- seq(0, 1, length = 10)
y <- 5 + x + (2 * (x^2)) + (-2 * (x^4))

dct_coefs <- dct(y)
recovered_y <- idct(dct_coefs)

plot(y, recovered_y)

}
