% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm_generic}
\alias{norm_generic}
\title{norm generic}
\usage{
norm_generic(
  .data,
  ...,
  .by = NULL,
  .norm_fun = identity_norm_fun,
  .by_formant = FALSE,
  .drop_orig = FALSE,
  .keep_params = FALSE,
  .names = "{.col}_n"
)
}
\description{
norm generic
}
