% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm_wattfab}
\alias{norm_wattfab}
\title{Watt & Fabricius method}
\usage{
norm_wattfab(
  .data,
  ...,
  .vowel_col,
  .high_front = NULL,
  .low = NULL,
  .by = NULL,
  .drop_orig = FALSE,
  .keep_params = FALSE,
  .names = "{.col}_wf"
)
}
\description{
Watt & Fabricius method
}
