% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidynorm-package.R
\docType{package}
\name{tidynorm-package}
\alias{tidynorm}
\alias{tidynorm-package}
\title{tidynorm: Tools for Tidy Vowel Normalization}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An implementation of tidy speaker vowel normalization. This includes generic functions for defining new normalization methods for points, formant tracks, and Discrete Cosine Transform coefficients, as well as convenience functions implementing established normalization methods. References for the implemented methods are: Johnson, Keith (2020) \doi{10.5334/labphon.196} Lobanov, Boris (1971) \doi{10.1121/1.1912396} Nearey, Terrance M. (1978) \url{https://sites.ualberta.ca/~tnearey/Nearey1978_compressed.pdf} Syrdal, Ann K., and Gopal, H. S. (1986) \doi{10.1121/1.393381} Watt, Dominic, and Fabricius, Anne (2002) \url{https://www.latl.leeds.ac.uk/article/evaluation-of-a-technique-for-improving-the-mapping-of-multiple-speakers-vowel-spaces-in-the-f1-f2-plane/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jofrhwld.github.io/tidynorm/}
  \item \url{https://github.com/JoFrhwld/tidynorm}
  \item Report bugs at \url{https://github.com/JoFrhwld/tidynorm/issues}
}

}
\author{
\strong{Maintainer}: Josef Fruehwald \email{JoFrhwld@gmail.com} [copyright holder]

}
\keyword{##}
\keyword{internal}
\keyword{namespace:}
\keyword{start}
\keyword{usethis}
